refactor: 💡 centralize leaf model prompts
