+++ apps/leaf/src/internal/agentRuntime/sessions/agentThreadTitle.ts
@@ -4,15 +4,13 @@ import { Agent } from "@mastra/core/agent";
 import { z } from "zod";
 import { DEFAULT_CHAT_ENV_MODEL } from "../../../lib/chatAgentConfig.js";
 import type { ChatDb } from "../../../lib/db.js";
+import { threadTitleInstructions } from "../../../prompts/threadTitlePrompt.js";
 import type { AgentThreadRef } from "../domain/agentTurnContext.js";
 import { setHarnessSessionTitleIfEmpty } from "../eve/repo.js";
 import { buildAgentThreadKey } from "./agentThreadKey.js";
 
 const titleSchema = z.strictObject({ title: z.string().min(1).max(60) });
 const TITLE_SOURCE_MAX_LENGTH = 2000;
-
-const TITLE_INSTRUCTIONS = `You title chat conversations for a billing dashboard.
-Given the user's first message, produce a short noun-phrase title (3-6 words, no quotes, no trailing punctuation) describing what they want. Examples: "Increase Pro plan price", "Set up usage billing", "List current plans".`;
 
 export const generateThreadTitle = async ({
 	logger,
@@ -25,7 +23,7 @@ export const generateThreadTitle = async ({
 		const agent = new Agent({
 			id: "leaf-thread-title",
 			name: "Leaf Thread Title",
-			instructions: TITLE_INSTRUCTIONS,
+			instructions: threadTitleInstructions,
 			model: DEFAULT_CHAT_ENV_MODEL,
 		});
 		const output = await agent.generate(
