+++ apps/leaf/src/prompts/threadTitlePrompt.ts
@@ -1,0 +1,2 @@
+export const threadTitleInstructions = `You title chat conversations for a billing dashboard.
+Given the user's first message, produce a short noun-phrase title (3-6 words, no quotes, no trailing punctuation) describing what they want. Examples: "Increase Pro plan price", "Set up usage billing", "List current plans".`;
