+++ apps/leaf/src/providers/slack/routing/classifySubscribedMessage.ts
@@ -6,25 +6,13 @@ import { z } from "zod";
 import type { AgentContextMessage } from "../../../internal/agentRuntime/domain/agentTurnContext.js";
 import { DEFAULT_SLACK_ROUTER_MODEL } from "../../../lib/chatAgentConfig.js";
 import { logger as rootLogger } from "../../../lib/logger.js";
+import { slackMessageRouterInstructions } from "../../../prompts/slackMessageRouterPrompt.js";
 
 const ROUTER_TIMEOUT_MS = ms.seconds(10);
 
 const routingSchema = z.strictObject({
 	disposition: z.enum(["respond", "ignore", "unsubscribe"]),
 });
-
-const instructions = `Classify the latest message in a Slack thread watched by Autumn Chat.
-
-Choose exactly one disposition:
-- respond: The message explicitly mentions Autumn Chat, continues or changes an Autumn billing, pricing, customer, plan, feature, or investigation request, answers the bot's question, or asks the bot for a new task.
-- ignore: The message is only an acknowledgment or social closure, is unrelated, or is side conversation for another person or bot.
-- unsubscribe: The message explicitly tells Autumn Chat or this bot to stop listening, leave the thread, or stop replying.
-
-Rules:
-- An explicit mention is respond unless the message explicitly asks the bot to stop.
-- A message that both acknowledges and requests Autumn work is respond.
-- Never unsubscribe merely because a message is unrelated.
-- When ambiguous, ignore.`;
 
 const applyMentionRouting = ({
 	disposition,
@@ -60,7 +48,7 @@ export const classifySubscribedMessage = async ({
 									latestMessage: { isMention, text },
 									recentMessages,
 								}),
-								system: instructions,
+								system: slackMessageRouterInstructions,
 							})
 						).output,
 					timeoutMessage: "Slack message routing timed out",
